[feat] mapview 그릴 때 rx 사용 (#31)
