+++ SpaceCapsule/SpaceCapsule/Scene/TabBar/CapsuleAccess/CapsuleDetail/CapsuleDetailViewController.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import RxSwift
-
 import SnapKit
 
 final class CapsuleDetailViewController: UIViewController, BaseViewController {
@@ -24,6 +23,10 @@ final class CapsuleDetailViewController: UIViewController, BaseViewController {
         viewModel?.addImage()
     
         bind()
+        
+        // TODO: 구현 후에 지우기
+        let center = GeoPoint(latitude: 37.583577, longitude: 127.019607)
+        viewModel?.fetchCapsuleMap(at: center, width: view.frame.width)
     }
     
     override func viewDidLayoutSubviews() {
@@ -39,11 +42,18 @@ final class CapsuleDetailViewController: UIViewController, BaseViewController {
     }
     
     func bind() {
-        viewModel?.input.imageData
+        viewModel?.output.imageData
             .withUnretained(self)
             .subscribe(onNext: { owner, items in
                 owner.mainView.imageCollectionView
                     .applySnapshot(items: items)
+            })
+            .disposed(by: disposeBag)
+        
+        viewModel?.output.mapSnapshot
+            .withUnretained(self)
+            .subscribe(onNext: { owner, mapImage in
+                owner.mainView.mapView.image = mapImage.first
             })
             .disposed(by: disposeBag)
     }
